QA-274 treats all egress traffic as tcp.
Libdssl infers traffic type as tcp or ssl based on tcp header.
As a result, if SSL traffic is found on a port configured for
tcp session, it still processes the traffic as ssl.
Due to the absence of ssl keys, the processing fails and ktrace
packet handler is not invoked.

Our usecase is that libdssl will only be used for tcp traffic
in egress direction. Egress ssl traffic will be handled by sslproxy.

This change is to process all egress traffic as tcp in libdssl,
regardless of traffic type inferred by libdssl.